+++ askbot/const/__init__.py
@@ -105,10 +105,6 @@ SELF_NOTIFY_EMAILED_POST_AUTHOR_WHEN_CHOICES = (
 #)
 
 REPLY_SEPARATOR_TEMPLATE = '==== %(user_action)s %(instruction)s -=-=='
-REPLY_WITH_COMMENT_TEMPLATE = _(
-    'Note: to reply with a comment, '
-    'please use <a href="mailto:%(addr)s?subject=%(subject)s">this link</a>'
-)
 REPLY_SEPARATOR_REGEX = re.compile(r'==== .* -=-==', re.MULTILINE|re.DOTALL)
 
 ANSWER_SORT_METHODS = (