+++ askbot/doc/source/changelog.rst
@@ -6,6 +6,7 @@ Development version (not released)
 * Added management command `askbot_rebuild_avatars`
 * Added option to choose gravatar or default avatar for new users
 * Message on the feedback page is editable
+* Added support for the NoCaptcha recaptcha
 
 0.7.53 (Apr 22, 2015)
 ---------------------